Move common operators and interfaces into templates.
