+++ sources/Maths/Maths/Vector3I.gen.cs
@@ -3,10 +3,18 @@ namespace Silk.NET.Maths
     using System.Collections;
     using System.Diagnostics.CodeAnalysis;
     using System.Numerics;
+    using System.Runtime.InteropServices;
+    using System.Text;
 
     partial struct Vector3I<T> :
         IEquatable<Vector3I<T>>,
-        IReadOnlyList<T>
+        IReadOnlyList<T>,
+        IFormattable,
+        IParsable<Vector3I<T>>,
+        ISpanFormattable,
+        ISpanParsable<Vector3I<T>>,
+        IUtf8SpanFormattable,
+        IUtf8SpanParsable<Vector3I<T>>
         where T : IBinaryInteger<T>
     {
         /// <summary>The X component of the vector.</summary>
@@ -23,6 +31,35 @@ namespace Silk.NET.Maths
 
         /// <summary>Initializes the vector with individual component values.</summary>
         public Vector3I(T x, T y, T z) => (X, Y, Z) = (x, y, z);
+
+        /// <summary>Initializes the vector from a span of 3 values.</summary>
+        public Vector3I(ReadOnlySpan<T> values)
+        {
+            if (values.Length != 3)
+                throw new ArgumentException("Input span must contain exactly 3 elements.", nameof(values));
+
+            X = values[0];
+            Y = values[1];
+            Z = values[2];
+        }
+
+        /// <summary>Gets a vector whose 3 elements are equal to one.</summary>
+        public static Vector3I<T> One => new(Scalar<T>.One);
+
+        /// <summary>Returns a vector whose 3 elements are equal to zero.</summary>
+        public static Vector3I<T> Zero => default;
+
+        /// <summary>Gets the vector (1, 0, 0).</summary>
+        public static Vector3I<T> UnitX => new(Scalar<T>.One, Scalar<T>.Zero, Scalar<T>.Zero);
+
+        /// <summary>Gets the vector (0, 1, 0).</summary>
+        public static Vector3I<T> UnitY => new(Scalar<T>.Zero, Scalar<T>.One, Scalar<T>.Zero);
+
+        /// <summary>Gets the vector (0, 0, 1).</summary>
+        public static Vector3I<T> UnitZ => new(Scalar<T>.Zero, Scalar<T>.Zero, Scalar<T>.One);
+
+        /// <summary>Gets a vector with all bits set for each component.</summary>
+        public static Vector3I<T> AllBitsSet => new Vector3I<T>(T.AllBitsSet, T.AllBitsSet, T.AllBitsSet);
 
         /// <inheritdoc/>
         T IReadOnlyList<T>.this[int index] => this[index];
@@ -61,6 +98,220 @@ namespace Silk.NET.Maths
             yield return Z;
         }
 
+        /// <summary>Copies the components of the vector to the specified array starting at index 0.</summary>
+        public void CopyTo(T[] array) => CopyTo(array, 0);
+
+        /// <summary>Copies the components of the vector to the specified array starting at the given index.</summary>
+        public void CopyTo(T[] array, int startIndex)
+        {
+            if (array == null)
+                throw new ArgumentNullException(nameof(array));
+            if (startIndex < 0 || startIndex + 3 > array.Length)
+                throw new ArgumentOutOfRangeException(nameof(startIndex));
+
+            array[startIndex] = X;
+            array[startIndex + 1] = Y;
+            array[startIndex + 2] = Z;
+        }
+
+        /// <summary>Copies the components of the vector to the specified span starting at index 0.</summary>
+        public void CopyTo(Span<T> span) => CopyTo(span, 0);
+
+        /// <summary>Copies the components of the vector to the specified span starting at the given index.</summary>
+        public void CopyTo(Span<T> span, int startIndex)
+        {
+            if (startIndex < 0 || startIndex + 3 > span.Length)
+                throw new ArgumentOutOfRangeException(nameof(startIndex));
+
+            span[startIndex] = X;
+            span[startIndex + 1] = Y;
+            span[startIndex + 2] = Z;
+        }
+
+        /// <summary>Returns a span over the vector components.</summary>
+        public Span<T> AsSpan() => MemoryMarshal.CreateSpan(ref X, 3);
+
+        /// <summary>Formats the vector as a string.</summary>
+        public override string ToString() =>
+            $"<{X}, {Y}, {Z}>";
+
+        /// <summary>Formats the vector as a string using the specified format and format provider.</summary>
+        public string ToString(string? format, IFormatProvider? formatProvider) =>
+            $"<{X.ToString(format, formatProvider)}, {Y.ToString(format, formatProvider)}, {Z.ToString(format, formatProvider)}>";
+
+        /// <summary>Parses a string to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static Vector3I<T> Parse(string s, IFormatProvider? provider) => Parse(s.AsSpan(), provider);
+
+        /// <summary>Parses a span to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static Vector3I<T> Parse(ReadOnlySpan<char> s, IFormatProvider? provider)
+        {
+            if (!TryParse(s, provider, out var result))
+                throw new FormatException("Invalid format for Vector3I.");
+
+            return result;
+        }
+
+        /// <summary>Formats the vector as a UTF-8 string using the specified format and format provider.</summary>
+        public bool TryFormat(Span<byte> utf8Destination, out int bytesWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
+        {
+            Span<char> xBuffer = stackalloc char[64];
+            Span<char> yBuffer = stackalloc char[64];
+            Span<char> zBuffer = stackalloc char[64];
+
+            if (!X.TryFormat(xBuffer, out int xChars, format, provider)||
+                !Y.TryFormat(yBuffer, out int yChars, format, provider)||
+                !Z.TryFormat(zBuffer, out int zChars, format, provider))
+            {
+                bytesWritten = 0;
+                return false;
+            }
+
+            int estimatedSize = Encoding.UTF8.GetByteCount(xBuffer[..xChars]) +
+                                Encoding.UTF8.GetByteCount(yBuffer[..yChars]) +
+                                Encoding.UTF8.GetByteCount(zBuffer[..zChars]) +
+                                Encoding.UTF8.GetByteCount("<, >");
+
+            if (utf8Destination.Length < estimatedSize)
+            {
+                bytesWritten = 0;
+                return false;
+            }
+
+            int totalBytes = 0;
+
+            totalBytes += Encoding.UTF8.GetBytes("<", utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(xBuffer[..xChars], utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(", ", utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(yBuffer[..yChars], utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(", ", utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(zBuffer[..zChars], utf8Destination[totalBytes..]);
+            totalBytes += Encoding.UTF8.GetBytes(">", utf8Destination[totalBytes..]);
+
+            bytesWritten = totalBytes;
+            return true;
+        }
+
+        /// <summary>Formats the vector as a string using the specified format and format provider.</summary>
+        public bool TryFormat(Span<char> destination, out int charsWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
+        {
+            Span<char> xBuffer = stackalloc char[64];
+            Span<char> yBuffer = stackalloc char[64];
+            Span<char> zBuffer = stackalloc char[64];
+
+            if (!X.TryFormat(xBuffer, out int xChars, format, provider) ||
+                !Y.TryFormat(yBuffer, out int yChars, format, provider) ||
+                !Z.TryFormat(zBuffer, out int zChars, format, provider))
+            {
+                charsWritten = 0;
+                return false;
+            }
+
+            int requiredLength = 1 + xChars + 2 + yChars + 2 + zChars + 1;
+
+            if (destination.Length < requiredLength)
+            {
+                charsWritten = 0;
+                return false;
+            }
+
+            int pos = 0;
+            destination[pos++] = '<';
+
+            xBuffer[..xChars].CopyTo(destination[pos..]);
+            pos += xChars;
+
+            destination[pos++] = ',';
+            destination[pos++] = ' ';
+
+            yBuffer[..yChars].CopyTo(destination[pos..]);
+            pos += yChars;
+
+            destination[pos++] = ',';
+            destination[pos++] = ' ';
+
+            zBuffer[..zChars].CopyTo(destination[pos..]);
+            pos += zChars;
+
+            destination[pos++] = '>';
+
+            charsWritten = pos;
+            return true;
+        }
+
+        /// <summary>Tries to parse a span to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static bool TryParse(ReadOnlySpan<char> s, IFormatProvider? provider, [MaybeNullWhen(false)] out Vector3I<T> result)
+        {
+            result = default;
+
+            s = s.Trim();
+            if (s.Length < 6 || s[0] != '<' || s[^1] != '>')
+                return false;
+
+            s = s[1..^1]; // Remove < and >
+
+            int commaX = s.IndexOf(',');
+            if (commaX < 0)
+                return false;
+
+            ReadOnlySpan<char> remainder1 = s.Slice(commaX + 1);
+            int commaYRelative = remainder1.IndexOf(',');
+            if (commaYRelative < 0)
+                return false;
+            int commaY = commaX + 1 + commaYRelative;
+
+            ReadOnlySpan<char> xSpan = s[..commaX].Trim();
+            ReadOnlySpan<char> ySpan = s[(commaX + 1)..commaY].Trim();
+            ReadOnlySpan<char> zSpan = s[(commaY + 1)..].Trim();
+
+            if (T.TryParse(xSpan, provider, out var x) &&
+                T.TryParse(ySpan, provider, out var y) &&
+                T.TryParse(zSpan, provider, out var z))
+            {
+                result = new Vector3I<T>(x, y, z);
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>Parses a UTF-8 span to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static Vector3I<T> Parse(ReadOnlySpan<byte> utf8Text, IFormatProvider? provider)
+        {
+            int charCount = Encoding.UTF8.GetCharCount(utf8Text);
+            Span<char> charBuffer = charCount <= 128 ? stackalloc char[charCount] : new char[charCount];
+            Encoding.UTF8.GetChars(utf8Text, charBuffer);
+            return Parse(charBuffer, provider);
+        }
+
+        /// <summary>Tries to parse a UTF-8 span to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static bool TryParse(ReadOnlySpan<byte> utf8Text, IFormatProvider? provider, [MaybeNullWhen(false)] out Vector3I<T> result)
+        {
+            int charCount = Encoding.UTF8.GetCharCount(utf8Text);
+            Span<char> charBuffer = charCount <= 128 ? stackalloc char[charCount] : new char[charCount];
+            Encoding.UTF8.GetChars(utf8Text, charBuffer);
+            return TryParse(charBuffer, provider, out result);
+        }
+
+        /// <summary>Tries to parse a string to a <see cref="Vector3I{T}"/> instance.</summary>
+        public static bool TryParse([NotNullWhen(true)] string? s, IFormatProvider? provider, [MaybeNullWhen(false)] out Vector3I<T> result) =>
+            TryParse(s.AsSpan(), provider, out result);
+
+        /// <summary>Parses a span to a <see cref="Vector3I{T}"/> instance.</summary>
+        static Vector3I<T> ISpanParsable<Vector3I<T>>.Parse(ReadOnlySpan<char> s, IFormatProvider? provider) =>
+            Parse(s, provider);
+
+        /// <summary>Parses a string to a <see cref="Vector3I{T}"/> instance.</summary>
+        static Vector3I<T> IParsable<Vector3I<T>>.Parse(string s, IFormatProvider? provider) =>
+            Parse(s, provider);
+
+        /// <summary>Tries to parse a span to a <see cref="Vector3I{T}"/> instance.</summary>
+        static bool ISpanParsable<Vector3I<T>>.TryParse(ReadOnlySpan<char> s, IFormatProvider? provider, [MaybeNullWhen(false)] out Vector3I<T> result) =>
+            TryParse(s, provider, out result);
+
+        /// <summary>Tries to parse a string to a <see cref="Vector3I{T}"/> instance.</summary>
+        static bool IParsable<Vector3I<T>>.TryParse([NotNullWhen(true)] string? s, IFormatProvider? provider, [MaybeNullWhen(false)] out Vector3I<T> result) =>
+            TryParse(s, provider, out result);
+
         /// <summary>Returns a boolean indicating whether the given two vectors are equal.</summary>
         /// <param name="left">The first vector to compare.</param>
         /// <param name="right">The second vector to compare.</param>
@@ -84,6 +335,75 @@ namespace Silk.NET.Maths
 
         /// <inheridoc/>
         public override int GetHashCode() => HashCode.Combine(X, Y, Z);
+
+        public static Vector3I<T> operator +(Vector3I<T> vector) =>
+            vector;
+
+        public static Vector3I<T> operator -(Vector3I<T> vector) =>
+            new Vector3I<T>(-vector.X, -vector.Y, -vector.Z);
+
+        public static Vector3I<T> operator +(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X + right.X, left.Y + right.Y, left.Z + right.Z);
+
+        public static Vector3I<T> operator -(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X - right.X, left.Y - right.Y, left.Z - right.Z);
+
+        public static Vector3I<T> operator *(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X * right.X, left.Y * right.Y, left.Z * right.Z);
+
+        public static Vector3I<T> operator /(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X / right.X, left.Y / right.Y, left.Z / right.Z);
+
+        public static Vector3I<T> operator %(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X % right.X, left.Y % right.Y, left.Z % right.Z);
+
+        public static Vector3I<T> operator +(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X + scalar, vector.Y + scalar, vector.Z + scalar);
+
+        public static Vector3I<T> operator -(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X - scalar, vector.Y - scalar, vector.Z - scalar);
+
+        public static Vector3I<T> operator *(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X * scalar, vector.Y * scalar, vector.Z * scalar);
+
+        public static Vector3I<T> operator *(T scalar, Vector3I<T> vector) =>
+            new Vector3I<T>(scalar * vector.X, scalar * vector.Y, scalar * vector.Z);
+
+        public static Vector3I<T> operator /(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X / scalar, vector.Y / scalar, vector.Z / scalar);
+
+        public static Vector3I<T> operator %(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X % scalar, vector.Y % scalar, vector.Z % scalar);
+
+        public static Vector3I<T> operator ~(Vector3I<T> vector) =>
+            new Vector3I<T>(~vector.X, ~vector.Y, ~vector.Z);
+
+        public static Vector3I<T> operator &(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X & right.X, left.Y & right.Y, left.Z & right.Z);
+
+        public static Vector3I<T> operator |(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X | right.X, left.Y | right.Y, left.Z | right.Z);
+
+        public static Vector3I<T> operator ^(Vector3I<T> left, Vector3I<T> right) =>
+            new Vector3I<T>(left.X ^ right.X, left.Y ^ right.Y, left.Z ^ right.Z);
+
+        public static Vector3I<T> operator &(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X & scalar, vector.Y & scalar, vector.Z & scalar);
+
+        public static Vector3I<T> operator &(T scalar, Vector3I<T> vector) =>
+            new Vector3I<T>(scalar & vector.X, scalar & vector.Y, scalar & vector.Z);
+
+        public static Vector3I<T> operator |(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X | scalar, vector.Y | scalar, vector.Z | scalar);
+
+        public static Vector3I<T> operator |(T scalar, Vector3I<T> vector) =>
+            new Vector3I<T>(scalar | vector.X, scalar | vector.Y, scalar | vector.Z);
+
+        public static Vector3I<T> operator ^(Vector3I<T> vector, T scalar) =>
+            new Vector3I<T>(vector.X ^ scalar, vector.Y ^ scalar, vector.Z ^ scalar);
+
+        public static Vector3I<T> operator ^(T scalar, Vector3I<T> vector) =>
+            new Vector3I<T>(scalar ^ vector.X, scalar ^ vector.Y, scalar ^ vector.Z);
     }
 
     static partial class Vector3I
